lua: add key_def lua module
There are two reasons to add this module:

* Incapsulate key_def creation from a Lua table (factor it out from
  merger's code).
* Support comparing tuple with tuple and/or tuple with key from Lua in
  the future.

The format of `parts` parameter in the `key_def.new(parts)` call is
compatible with the following structures:

* box.space[...].index[...].parts;
* net_box_conn.space[...].index[...].parts.

Needed for #3276.
Needed for #3398.

@TarantoolBot document
Title: Document built-in key_def lua module

Now there is only stub with the `key_def.new(parts)` function that
returns cdata<struct key_def &>. The only way to use it for now is pass
it to the merger.

This module will be improved in the scope of
https://github.com/tarantool/tarantool/issues/3398

See the commit message for more info.