Transaction support for applier
Applier fetch incoming rows to form a transaction and then apply it.
In case of  replication all local changes moved to an journal entry
tail to form a separate transaction (like autonomous transaction)
to be able to replicate changes back. Applier assumes that transactions
could not be mixed in a replication stream.

Closes: #2798
Needed for: #980