sql: rework sqlSrcListEnlarge to set diag
Refactored sqlSrcListEnlarge routine as sql_src_list_enlarge and
reworked to use diag_set in case of memory allocation error. This
will ensure that the sqlSrcListAppend function throws an error
using diag in subsequent patches.

This patch refers to a series of preparatory patches that provide
the use of Tarantool errors in the call tree that includes
sqlNormalizeName, since this call can later return errors.

This patch is not self-sufficient, its usage in sqlSrcListAppend
remained to be non-Tarantool (for now). It means, that if
sql_src_list_enlarge fails in sqlSrcListAppend the diag will
never be thrown.

Part of #3931