sql: replace instr() with position()
Before this patch we had instr() SQL function.
After the patch it is renamed to position().
Also a few things have been changed: arguments
order, allowed arguments types and usage of
collations.

The order of the arguments has changed:
Before: instr(haystack, needle).
After: position(needle, haystack).

Type checking became more strict: before it was
possible to call the function with INTEGER arguments
or with arguments of different types. Now both arguments
must have the same type and be either text or binary
strings.

Before the patch collations were not taken into
consideration during the search. Now it is fixed, and
both implicit (column) collations and explicit
(using COLLATE expression) are used. There are
several possible cases:

- One of collations is mentioned alongside with explicit
COLLATE clause, which forces this collation over another
one. It is allowed to have the same forced collations;
- Both collations are derived from table columns and they
are the same;
- One collation is derived from table column and another
one is not specified (i.e. COLL_NONE);

In all other cases they are not accounted to be compatible
and error is raised.

Related to #3933

@TarantoolBot document
Title: instr() is replaced with position()
Changes are described in the associated commit
message.