sio: make sio_strfaddr taking const struct sockaddr
SWIM module API is going to provide a set of clear and pure
functions with appropriately settled const qualifiers. And it
wants to use sio_strfaddr() to provide to user an easy way to
get a pointer to URI of a SWIM member stored in a const memory.

It requires this two-line modification of sio module.