Make memtx engine responsible for a ddl transaction tracking
Allow one ddl to be between begin and commit/rollback. This is essential
for transactional replication. This patch introduces a sequence of
changes and refactorings:
 * setup on_yield/on_stop trigger despite of transaction auto commit behavior
 * reset on_yield trigger in case of ddl operation
 * make memtx engine responsible for ddl single statement restriction
 * get rid of check_single_statement function
 * add some tests

Note: not any box.schema operation is capable to be between begin/commit
yet because of the fact that it could involve more than one ddl-space.